% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRaWR.R
\name{threeCol2MaxMat}
\alias{threeCol2MaxMat}
\title{threeCol2MaxMat}
\usage{
threeCol2MaxMat(a = c("a", "b", "c", "c"), b = c("a", "b", "b", "b"),
  v = c(1, 2, 3, 4))
}
\arguments{
\item{a}{(vector): vector of source node names.}

\item{b}{(vector): vector of target node names.}

\item{v}{(vector): vector of edge weights names.}
}
\value{
sparce Matrix.
}
\description{
This function takes a three vectors of equal length (source nodes, target nodes, and edge weights) and return the adjacency matrix as a sparse Matrix.
}
\examples{
threeCol2MaxMat(a = c("a","b","c","c"), b = c("a","b","b","b"), v = c(1,2,3,4))
}
\keyword{matrix}
\keyword{sparce}

