% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRaWR.R
\docType{data}
\name{transmat}
\alias{transmat}
\title{Sample transition matrix.}
\format{a Matrix containing the normalized transition matrix from the test network.}
\usage{
transmat
}
\description{
A Matrix containing tthe normalized transition matrix from the test network
}
\keyword{datasets}

